pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wanamaker"
version = "0.0.0"
description = "Open-source Bayesian marketing mix model for mid-market teams. Knowing which half."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Wanamaker contributors" }]
keywords = ["mmm", "marketing-mix-modeling", "bayesian", "marketing-analytics"]
classifiers = [
    "Development Status :: 1 - Planning",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

# Production dependencies.
#
# Deliberately omitted (pending Phase -1 engine decision per AGENTS.md):
#   - pymc, numpyro, jax, cmdstanpy
# Deliberately omitted (per Hard Rules 1 & 2 — no network, no LLM):
#   - requests, httpx, urllib3 (beyond what pip pulls in transitively)
#   - openai, anthropic, or any LLM SDK
dependencies = [
    "numpy>=1.26",
    "pandas>=2.2",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "typer>=0.12",
    "jinja2>=3.1",
    "matplotlib>=3.8",
    "seaborn>=0.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
# Quick-mode preview / Bayesian cross-check only. Never used for ROI/saturation/adstock.
# Hard Rule 3 in AGENTS.md.
xgboost = ["xgboost>=2.0"]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
]

[project.scripts]
wanamaker = "wanamaker.cli:app"

[project.urls]
Documentation = "https://wanamaker.dev"  # placeholder
Source = "https://github.com/wanamaker/wanamaker"  # placeholder

[tool.hatch.build.targets.wheel]
packages = ["src/wanamaker"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", "F", "W",  # pycodestyle, pyflakes
    "I",            # isort
    "N",            # pep8-naming
    "UP",           # pyupgrade
    "B",            # bugbear
    "SIM",          # simplify
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra --strict-markers"
markers = [
    "benchmark: benchmark-driven acceptance tests (slow)",
    "integration: end-to-end integration tests",
    "engine: requires a configured Bayesian engine (skipped pre-Phase -1)",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true
